sidecar/pyproject.toml
[project]
name = "eps-sidecar"
version = "0.1.0"
description = "Spawned Claude Code agent over HTTP/SSE for the EPS Dashboard chat rail."
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.7",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "anyio>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eps_sidecar"]
